flask_migration: move webpy/flask specific files in separate folders
